fix: exhaustive bug fixes across auth, security, templates and PDFs
Auth & sessions:
- H-02: User.get_id() binds session to password_hash[-16:]; user_loader validates suffix
- H-01/C-13: user_loader checks is_active; SECRET_KEY raises RuntimeError in production

Security:
- C-01/C-02: always anonymize users (never hard-delete); random suffix prevents pre-registration IDOR
- H-06: validate hex color format before saving org config (prevents ReportLab crash)
- C-07: server-side org validation for producto_id in gestionar_inventario_almacen
- H-18: @check_permission('perm_view_gastos') on all 8 service routes
- H-19: @check_org_permission + rate limit on AI api endpoints
- H-23: push subscribe filters by user_id and refreshes org_id

Data integrity:
- C-04: remove organizacion_id from Stock() constructor (column does not exist)
- C-05: Presupuesto.monto Float→Numeric(10,2); added to fix-float CLI command
- C-06: server-side stock re-verify before sin_stock warehouse deletion
- C-08: editar_producto creates Movimiento for stock changes instead of direct overwrite
- H-04: AuditLog.descripcion truncated to 500 chars to prevent StringDataRightTruncation
- H-16: marcar_pago_pagado early-return guard if pago already pagado (prevents duplicate Gasto)

PDF/ReportLab:
- H-13/H-14: _xml_escape applied to all user-data in Paragraph() in inventory + reports blueprints
- Added None guard for salida.almacen in salida PDF

Templates/XSS:
- C-09: window.escHtml exported from base.html IIFE for global access
- C-10: almacenes.html error flag prevents destructive sin_stock mode on network failure
- C-11: server + client validation for transferir_separado missing destination
- C-12: dashboard search suggestions use DOM API (no innerHTML)
- H-20/H-21/H-22: _esc() helpers added in all affected template IIFEs; user data escaped
- H-24: exportar_gastos_excel date range clamped to 24 months max

Other:
- M-20: mes clamped to 1–12 in historial_salidas (prevents calendar.IllegalMonthError)
- M-22: removed duplicate CSRF token from login form (form.hidden_tag() already emits one)
- M-23: username apostrophe escaped in abrirModalPassword onclick
- H-03: log_actividad called before db.session.commit() in api_sync

+++ app/models/finance.py
@@ -133,7 +133,7 @@ class Presupuesto(db.Model):
     categoria       = db.Column(db.String(50), nullable=False)
     anio            = db.Column(db.Integer,    nullable=False)
     mes             = db.Column(db.Integer,    nullable=True)
-    monto           = db.Column(db.Float,      nullable=False)
+    monto           = db.Column(db.Numeric(10, 2), nullable=False)
     organizacion_id = db.Column(db.Integer,    db.ForeignKey('organizacion.id'), nullable=False)
     creado_en       = db.Column(db.DateTime,   default=now_mx)
